refactor(cli): move policy-prompt module into commands, delete internal/ui
The interactive policy-prompt functions (PromptForPolicies,
PromptForPoliciesUpdate, ParseCapabilities) lived in internal/ui — an
infrastructure package doing terminal I/O and importing auth/domain, yet
used only by the create-client and update-client CLI commands.

Relocate them to cmd/app/commands (package commands) beside their only
callers and delete the now-empty internal/ui package. Pure move: same
logic, same tests, same behaviour. The seam now lives where the
behaviour does, and the infra→domain import direction is gone.
